==> 3.3_eval_pymorphy_seq2seq_OpenCorp.ipynb <==
from keras.models import Model, model_from_json, load_model
from keras.layers import Input, LSTM, Dense
from keras.callbacks import ModelCheckpoint
import numpy as np
import keras
from tqdm import tqdm
import pandas as pd
import pymorphy2
import matplotlib.pyplot as plt
import seaborn as sns



### data

path_to_rfile = 'OpenCorpora_Texts/unamb_sent_14_6.conllu'
path_to_wfile = 'data/eval/eval_OpenCorpora.csv'

#для OpenCorpora
data = pd.read_csv(path_to_rfile, encoding='utf-8',sep='\t',names = ['index','form','lemma','POS','1','gram','2','3','4','5'])

data.shape

data = data.loc[data['POS']=='NOUN']
data = data[['lemma','gram','form']]

data.shape

def converter(gram):
    grams = gram.split('|')
    dc_gram = {}
    for gram in grams:
        k,v = gram.split('=')
        dc_gram[k] = v
    tag = ''
    if dc_gram['Number'].startswith('S'):
        tag += '1'
    elif dc_gram['Number'].startswith('P'):
        tag += '2'
    else: print(gram)
    tag += dc_gram['Case'][0]
    return tag

for i in range(0,len(data)):
    try:
        data.iloc[i,1] = converter(data.iloc[i,1])
    except Exception as e:
        #print(e)
        #print(data.iloc[[i]])
        data.iloc[[i]] = 0
len(data)   

data = data[data.gram != 0]
len(data)

data['lemma'] = data.apply(lambda row: row['lemma'].lower(), axis=1)

data['form'] = data.apply(lambda row: row['form'].lower(), axis=1)

data['x'] = data.apply(lambda row: row['lemma'] + row['gram'], axis=1)

#data = data[['x','form']]

### checking pymorphy

morph = pymorphy2.MorphAnalyzer()

word = morph.parse('зож')

word

#table from seq2seq to pymorphy
trans_number = {'1':'sing', '2':'plur'}
trans_case = {'N':'nomn','G':'gent','D':'datv','A':'accs','I':'ablt','L':'loct'}

lemma = "твиттер"
number = 'sing'
case = 'loct'

py_lemmas = morph.parse(lemma)
py_lemma = '-'
for p_lemma in py_lemmas:
    if p_lemma.tag.POS == 'NOUN':
        py_lemma = p_lemma
        break 

py_lemma

py_method = str(py_lemma.methods_stack[0][0])
py_method

prediction = py_lemma.inflect({number, case}).word
prediction

### checking seq2seq model

from model import predict_form, predict_paradigm

enc = load_model('models/encoder_total.h5')
dec = load_model('models/decoder_total.h5')

predict_paradigm('зашквар',enc,dec)

data.iloc[[35]]

### looping through data and predicting forms with pymorphy

%%time
pymorphy_predictions = []
pymorphy_eval = []
pymorphy_exeptions = []
pymorphy_methods = []
for row in range(len(data)):
    lemma = data.iloc[row,0]
    #print(row)
    #print(data.iloc[row,1])
    number,case = str(data.iloc[row,1])
    number = trans_number[number]
    case = trans_case[case]
    py_lemmas = morph.parse(lemma)
    py_lemma = '-'
    for p_lemma in py_lemmas:
        if p_lemma.tag.POS == 'NOUN':
            py_lemma = p_lemma
            break 
    try:
        py_method = str(py_lemma.methods_stack[0][0])
        prediction = py_lemma.inflect({number, case}).word        
    except Exception as e:
        pymorphy_exeptions.append((number,case,lemma))
        prediction =  '-'
        py_method = '-'      
        #print('{} {}({})'.format(number,case,lemma))  
    pymorphy_methods.append(py_method)
    pymorphy_predictions.append(prediction)
    true_form = data.iloc[row,2]
    if prediction == true_form:
        pymorphy_eval.append('True')
    else:
        pymorphy_eval.append('False')

data['pymorphy_predictions'] = pymorphy_predictions
data['pymorphy_eval'] = pymorphy_eval
data['pymorphy_methods'] = pymorphy_methods

### predicting forms with seq2seq

%%time
seq2seq_predictions = []
seq2seq_eval = []
for row in tqdm(range(len(data))):
#for row in tqdm(range(100)):
    inp = data.iloc[row,3]
    out = predict_form(inp,enc,dec).rstrip()
    seq2seq_predictions.append(out)
    if out == data.iloc[row,2]:
        seq2seq_eval.append('True')
    else:
        seq2seq_eval.append('False')

data['seq2seq_predictions'] = seq2seq_predictions
data['seq2seq_eval'] = seq2seq_eval


#uploading results from file
seq2seq_predictions = open('testing/model_predict_full_letters.txt','r',encoding = 'utf-8').read().split('\n')
data['seq2seq_predictions'] = seq2seq_predictions
seq2seq_eval = []
for row in range(data.shape[0]):
    true_form = data.iloc[row,4]
    prediction = data.iloc[row,7]
    if prediction == true_form:
        seq2seq_eval.append('True')
    else:
        seq2seq_eval.append('False')
data['seq2seq_eval'] = seq2seq_eval      

data.to_csv(path_to_wfile, encoding='utf-8',sep='\t', index=False)
#writer = pd.ExcelWriter(path_to_wfile.replace('.csv', '.xlsx'))
#data.to_excel(writer,'Sheet1')
#writer.save()
#для замеров аккуратности без учета "ошибок" на неизменяемых, pl.tantum, sg.tantum
#data_filtered = data[data.apply(lambda x: not x['classtag'].endswith(('-','+','0')), axis=1)] 
#data_filtered.to_csv(path_to_wfile_filtered, encoding='utf-8',sep='\t', index=False)

data = pd.read_csv(path_to_wfile, encoding='utf-8',sep='\t')

data['pymorphy_predictions'] = data['pymorphy_predictions'].apply(lambda x: x.replace("ё","е"))

data['seq2seq_predictions'] = data['seq2seq_predictions'].apply(lambda x: x.replace("ё","е"))

# не словарные слова для pymorphy
data_filtered = data[data.apply(lambda x: x['pymorphy_methods'] != '<DictionaryAnalyzer>', axis=1)]

len(data)

len(data_filtered)

data_filtered

data_filtered[(data_filtered['pymorphy_eval'] == False) & (data_filtered['seq2seq_eval'] == True)]

len(data_filtered[(data_filtered['seq2seq_eval'] == False)])

### accuracy

def accuracy(model,data):
    column = model + '_eval'
    eval_ = data[column]
    value_counts =  eval_.value_counts()
    print(value_counts)
    print((len(eval_)))
    acc = value_counts[True]/(len(eval_))
    acc = round(acc,4)
    print('for model {}:'.format(model))
    print('False:', value_counts[False])
    print('True:', value_counts[True])
    print('accuracy:',acc)

accuracy('seq2seq', data)

accuracy('pymorphy', data)

accuracy('seq2seq', data_filtered)

accuracy('pymorphy', data_filtered)
