==> 4.2_eval_retraining_models.ipynb <==
import warnings
warnings.filterwarnings('ignore')

from keras.models import Model, model_from_json, load_model
from keras.layers import Input, LSTM, Dense
from keras.callbacks import ModelCheckpoint
import numpy as np
import keras
from tqdm import tqdm
import pandas as pd
from time import time

from model import predict_form, predict_paradigm

### model try

enc = load_model('retraining_models/encoder-13.h5')
dec = load_model('retraining_models/decoder-13.h5')

predict_form('кукурузохранилище1I',enc,dec)

predict_paradigm('кукурузохранилище',enc,dec)

### accuracy

def accuracy(data):
    eval_ = data['seq2seq_eval']
    value_counts =  eval_.value_counts()
    try:
        truth = value_counts[True]
    except:
        truth = 0
    try:
        errors = value_counts[False]
    except:
        errors = 0
    acc = truth/(len(eval_))
    acc = round(acc,4)
    print('False:', errors)
    print('True:', truth)
    print('accuracy:',acc)
    print('\n')
    return acc

### data

path_to_rfile = 'data/test/test_full.csv' 
data_full = pd.read_csv(path_to_rfile, encoding='utf-8',sep='\t')

data = data_full[data_full.apply(lambda x: not x['classtag'].endswith(('-','+','0')), axis=1)]

len(data)

for i in range(19,39):
    print('evaluating model_{}...'.format(i))
    wfile = 'training_loop/eval/eval_model_{}.csv'.format(i)
    data_eval = data.copy()
    enc_name = 'retraining_models/encoder-{}.h5'.format(i)
    dec_name = 'retraining_models/decoder-{}.h5'.format(i)
    enc = load_model(enc_name)
    dec = load_model(dec_name)
    seq2seq_predictions = []
    seq2seq_eval = []
    t = time()
    for row in tqdm(range(data.shape[0])):
        inp = data_eval.iloc[row,0]+str(data_eval.iloc[row,2])
        out = predict_form(inp,enc,dec).rstrip()
        seq2seq_predictions.append(out)
        if out == data_eval.iloc[row,4]:
            seq2seq_eval.append('True')
        else:
            seq2seq_eval.append('False')
    data_eval['seq2seq_predictions'] = seq2seq_predictions
    data_eval['seq2seq_eval'] = seq2seq_eval
    print('evaluating took {} min'.format(round((time() - t)/60,4)))
    print('for model_{}:'.format(i))
    acc = accuracy(data_eval)
    data_eval.to_csv(wfile, encoding='utf-8',sep='\t', index=False)
    with open('training_loop/eval/eval_result.csv','a',encoding = 'utf-8') as file:
        file.write('{}\t{}\n'.format(i,acc))

###  по уже прогнанному тесту

def accuracy(data):
    a,t,f = 0,0,0
    eval_ = data['seq2seq_eval']
    values =  eval_.value_counts()
    t = values[True]
    f = values[False]
    a = t/(len(eval_))
    a = round(a,4)
    print('Total count:', len(eval_))
    print('False:', f)
    print('True:', t)
    print('accuracy:',a)
    print('\n')
    return a

with open('training_loop/eval/eval_true_result.csv','w',encoding = 'utf-8') as file:
    for i in range(1,40):
        print('evaluating model_{}...'.format(i))
        wfile = 'training_loop/eval/eval_model_{}.csv'.format(i)
        data_eval = pd.read_csv(wfile, encoding='utf-8',sep='\t')
        acc = accuracy(data_eval)
        #data_eval.to_csv(wfile, encoding='utf-8',sep='\t', index=False)
        file.write('{}\t{}\n'.format(i,acc))
